src/main/java/com/liyosi/fruitshop/controllers/v1/CategoryController.java
package com.liyosi.fruitshop.controllers.v1;

import com.liyosi.fruitshop.api.v1.model.CategoryDTO;
import com.liyosi.fruitshop.api.v1.model.CategoryListDTO;
import com.liyosi.fruitshop.services.CategoryService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;

/**
 * Created by liyosi on Sep, 2018
 */
@Controller
@RequestMapping("/api/v1/categories/")
public class CategoryController {

  private final CategoryService categoryService;

  public CategoryController(CategoryService categoryService) {
    this.categoryService = categoryService;
  }

  @GetMapping
  public ResponseEntity<CategoryListDTO> getAllCategories() {
    return new ResponseEntity<>(new CategoryListDTO(categoryService.getAllCategories()), HttpStatus.OK);
  }

  @GetMapping("{name}")
  public ResponseEntity<CategoryDTO> getCategory(@PathVariable String name) {
    return new ResponseEntity<CategoryDTO>(categoryService.getCategoryByName(name), HttpStatus.OK);
  }
}
